move the function file of sendEmail to lib folder
