+++ backend/lib/sendEmail.js
@@ -1,0 +1,14 @@
+const nodemailer = require("nodemailer");
+
+let smtpTransport = nodemailer.createTransport({
+  host: "smtp.gmail.com",
+  auth: {
+    type: "OAuth2",
+    user: "spiritxhx@gmail.com",
+    clientId: process.env.GOOGLE_CLIENT_ID,
+    clientSecret: process.env.GOOGLE_CLIENT_SECRET,
+    refreshToken: process.env.GOOGLE_REFRESH_TOKEN
+  }
+});
+
+module.exports = {smtpTransport};
